refactor: move YandexCalendar-specific code into YandexCalendar/ subfolders
