+++ src/Pet.Jira.Application/Extensions/YandexCalendar/Dto/YandexCalendarSettingsDto.cs
@@ -1,4 +1,4 @@
-namespace Pet.Jira.Application.Extensions.Dto
+namespace Pet.Jira.Application.Extensions.YandexCalendar.Dto
 {
     public record YandexCalendarSettingsDto(string Login, string AppPassword);
 }
