+++ src/Pet.Jira.Application/Extensions/YandexCalendar/ICalendarService.cs
@@ -1,10 +1,10 @@
-using Pet.Jira.Application.Extensions.Dto;
+using Pet.Jira.Application.Extensions.YandexCalendar.Dto;
 using System;
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Pet.Jira.Application.Extensions
+namespace Pet.Jira.Application.Extensions.YandexCalendar
 {
     public record CalendarCredentials(string Login, string AppPassword);
