+++ src/Pet.Jira.Application/Extensions/YandexCalendar/Queries/GetYandexCalendarSettings.cs
@@ -1,9 +1,9 @@
 using MediatR;
-using Pet.Jira.Application.Extensions.Dto;
+using Pet.Jira.Application.Extensions.YandexCalendar.Dto;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Pet.Jira.Application.Extensions.Queries
+namespace Pet.Jira.Application.Extensions.YandexCalendar.Queries
 {
     public class GetYandexCalendarSettings
     {
