+++ src/Pet.Jira.Infrastructure/Extensions/YandexCalendar/YandexCalendarSettings.cs
@@ -1,4 +1,4 @@
-namespace Pet.Jira.Infrastructure.Extensions
+namespace Pet.Jira.Infrastructure.Extensions.YandexCalendar
 {
     // JSON stored in UserExtension.Settings — password is encrypted
     internal record YandexCalendarSettings(string Login, string AppPasswordEncrypted);
