+++ src/Pet.Jira.Web/Components/Extensions/ExtensionsPage.razor
@@ -1,4 +1,5 @@
 @using Pet.Jira.Web.Components.Extensions
+@using Pet.Jira.Web.Components.Extensions.YandexCalendar
 
 <div class="extv-glow">
     <div class="extv-glow__head">
