+++ src/Pet.Jira.Web/Components/Extensions/YandexCalendar/YandexCalendarExtensionCard.razor.cs
@@ -2,12 +2,12 @@ using MediatR;
 using Microsoft.AspNetCore.Components;
 using MudBlazor;
 using Pet.Jira.Application.Authentication;
-using Pet.Jira.Application.Extensions.Commands;
-using Pet.Jira.Application.Extensions.Dto;
-using Pet.Jira.Application.Extensions.Queries;
+using Pet.Jira.Application.Extensions.YandexCalendar.Commands;
+using Pet.Jira.Application.Extensions.YandexCalendar.Dto;
+using Pet.Jira.Application.Extensions.YandexCalendar.Queries;
 using System.Threading.Tasks;
 
-namespace Pet.Jira.Web.Components.Extensions
+namespace Pet.Jira.Web.Components.Extensions.YandexCalendar
 {
     public partial class YandexCalendarExtensionCard : ComponentBase
     {
