+++ src/Pet.Jira.Web/Components/Worklogs/WorklogDay.razor
@@ -1,5 +1,5 @@
 @using Pet.Jira.Web.Common
-@using Pet.Jira.Application.Extensions.Dto
+@using Pet.Jira.Application.Extensions.YandexCalendar.Dto
 @using MudBlazor
 
 <tr class="pet-table-summary-tr">
