+++ src/Pet.Jira.Web/Components/Worklogs/WorklogDay.razor.cs
@@ -2,8 +2,8 @@ using MediatR;
 using Microsoft.AspNetCore.Components;
 using MudBlazor;
 using Pet.Jira.Application.Authentication;
-using Pet.Jira.Application.Extensions.Dto;
-using Pet.Jira.Application.Extensions.Queries;
+using Pet.Jira.Application.Extensions.YandexCalendar.Dto;
+using Pet.Jira.Application.Extensions.YandexCalendar.Queries;
 using Pet.Jira.Application.Worklogs.Commands;
 using Pet.Jira.Application.Worklogs.Dto;
 using Pet.Jira.Web.Shared;
