+++ tests/Pet.Jira.UnitTests/Application/Extensions/YandexCalendar/GetCalendarEventsHandlerTests.cs
@@ -1,14 +1,15 @@
 using Moq;
 using NUnit.Framework;
 using Pet.Jira.Application.Extensions;
-using Pet.Jira.Application.Extensions.Dto;
-using Pet.Jira.Application.Extensions.Queries;
+using Pet.Jira.Application.Extensions.YandexCalendar;
+using Pet.Jira.Application.Extensions.YandexCalendar.Dto;
+using Pet.Jira.Application.Extensions.YandexCalendar.Queries;
 using System;
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Pet.Jira.UnitTests.Application.Extensions
+namespace Pet.Jira.UnitTests.Application.Extensions.YandexCalendar
 {
     [TestFixture]
     public class GetCalendarEventsHandlerTests
