+++ tests/Pet.Jira.UnitTests/Application/Extensions/YandexCalendar/UpsertExtensionHandlerTests.cs
@@ -1,8 +1,8 @@
 using Moq;
 using NUnit.Framework;
 using Pet.Jira.Application.Extensions;
-using Pet.Jira.Application.Extensions.Commands;
-using Pet.Jira.Application.Extensions.Dto;
+using Pet.Jira.Application.Extensions.YandexCalendar.Commands;
+using Pet.Jira.Application.Extensions.YandexCalendar.Dto;
 using Pet.Jira.Application.Security;
 using Pet.Jira.Domain.Entities.Extensions;
 using System;
@@ -10,7 +10,7 @@ using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Pet.Jira.UnitTests.Application.Extensions
+namespace Pet.Jira.UnitTests.Application.Extensions.YandexCalendar
 {
     [TestFixture]
     public class UpsertExtensionHandlerTests
